refactor(picker, taro): modify to gen better dts
